Create Conekta Oxxo payment backend
We created the controller to handle requests from the TiregetMX store
and the Conekta API and also added the Payment Method model to be
displayed in the admin Payment Methods section.

The payment source in the models directory allows the orders state
to be changed to completed and move on to the next step in the
checkout process.